Astar algorithm

+++ src/main/java/FileOperator.java
@@ -8,19 +8,19 @@ import java.util.Arrays;
 import java.util.List;
 
 public class FileOperator {
-    public static int[][] readBoard(Path p) throws IOException {
+    public static byte[][] readBoard(Path p) throws IOException {
         if (Files.exists(p)) {
             List<String> lines = Files.readAllLines(p);
             List<Integer> args = new ArrayList<>();
             for (String arg : lines.remove(0).split("\\s+")) {
                 args.add(Integer.parseInt(arg));
             }
-            int[][] board = new int[args.get(0)][args.get(1)];
+            byte[][] board = new byte[args.get(0)][args.get(1)];
             int i = 0;
             int j = 0;
             for (String s : lines) {
                 for (String arg : s.split("\\s+")) {
-                    board[i][j] = Integer.parseInt(arg);
+                    board[i][j] = Byte.parseByte(arg);
                     j++;
                 }
                 j = 0;
